# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Got Journey Thailand API",
    "description": "Travel Booking Platform API - Hotels, Cars, and Package Bookings",
    "version": "1.0.0",
    "contact": {
      "name": "Got Journey Thailand",
      "email": "support@gotjourneythailand.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development Server"
    },
    {
      "url": "https://api.gotjourneythailand.com",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/api/wishlist": {
      "get": {
        "tags": ["Wishlist"],
        "summary": "Get user's wishlist (requires auth)",
        "security": [{ "UserAuth": [] }],
        "responses": {
          "200": { "description": "Successful response" }
        }
      },
      "post": {
        "tags": ["Wishlist"],
        "summary": "Add to wishlist (requires auth)",
        "security": [{ "UserAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/WishlistInput" }
            }
          }
        },
        "responses": {
          "201": { "description": "Added to wishlist" }
        }
      },
      "delete": {
        "tags": ["Wishlist"],
        "summary": "Remove from wishlist (requires auth)",
        "security": [{ "UserAuth": [] }],
        "parameters": [
          { "name": "item_type", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "item_id", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Removed from wishlist" }
        }
      }
    },
    "/api/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Search hotels and cars",
        "parameters": [
          { "name": "type", "in": "query", "schema": { "type": "string", "enum": ["hotel", "car", "all"] } },
          { "name": "query", "in": "query", "schema": { "type": "string" } },
          { "name": "min_price", "in": "query", "schema": { "type": "number" } },
          { "name": "max_price", "in": "query", "schema": { "type": "number" } },
          { "name": "location", "in": "query", "schema": { "type": "string" } },
          { "name": "rating", "in": "query", "schema": { "type": "number" } },
          { "name": "guests", "in": "query", "schema": { "type": "number" } },
          { "name": "sort_by", "in": "query", "schema": { "type": "string", "enum": ["price_asc", "price_desc", "rating", "newest"] } },
          { "name": "lang", "in": "query", "schema": { "type": "string", "enum": ["th", "en"] } }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "results": { "type": "array" },
                    "hotels": { "type": "integer" },
                    "cars": { "type": "integer" },
                    "total": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_token"
      },
      "UserAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_token"
      }
    },
    "schemas": {
...
      "WishlistInput": {
        "type": "object",
        "required": ["item_type", "item_id"],
        "properties": {
          "item_type": { "type": "string", "enum": ["hotel", "car"] },
          "item_id": { "type": "string" }
        }
      }
    }
  }
}